using same rendering logic for top level attachments
